backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
python-jose[cryptography]
python-multipart 